VM and repo hu5 and hu6

+++ app/src/main/java/com/cristian/miniproyecto2/view/RecyclerAdapter.kt
@@ -36,7 +36,7 @@ class RecyclerAdapter(private val listaArticulos: MutableList<Articulo>): Recycl
             val bundle = Bundle().apply {
                 putLong("idArticulo", idArticulo)
                 putString("nameArticulo", nameArticulo)
-                putLong("priceArticulo", priceArticulo)
+                putDouble("priceArticulo", priceArticulo.toDouble())
                 putLong("quantityArticulo", quantityArticulo)
             }
